+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/dialogs/TableAssetManagerDialogController.java
@@ -181,6 +181,9 @@ public class TableAssetManagerDialogController implements Initializable, DialogC
   private Button assetSourceBtn;
 
   @FXML
+  private Button searchBtn;
+
+  @FXML
   private Button addAudioBlank;
 
   @FXML
@@ -1193,6 +1196,13 @@ public class TableAssetManagerDialogController implements Initializable, DialogC
 
     this.setDefaultBtn.setVisible(true);
     this.renameBtn.setVisible(true);
+
+    boolean mediaSearchSet = !StringUtils.isEmpty(game.getMediaSearch());
+    searchField.setDisable(mediaSearchSet);
+    assetSourceComboBox.setDisable(mediaSearchSet);
+    assetSourceBtn.setDisable(mediaSearchSet);
+    clearCacheBtn.setDisable(mediaSearchSet);
+    searchBtn.setDisable(mediaSearchSet);
 
     if (!isEmbeddedMode()) {
       nextButton.setDisable(false);